requirements.txt
click==8.0.3
Flask==2.0.2
future==0.18.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-whois==0.7.3
Werkzeug==2.0.2
